nfs3: move nfs3 specific handling into own file
