+++ rust/src/nfs/mod.rs
@@ -21,8 +21,9 @@ pub mod nfs_records;
 pub mod nfs2_records;
 pub mod nfs3_records;
 pub mod nfs4_records;
-pub mod nfs4;
 pub mod nfs;
+pub mod nfs3;
+pub mod nfs4;
 pub mod log;
 
 //#[cfg(feature = "lua")]
